codegen/cpu.rs for CpuCodeGenerator
